remove Vector1
